manage a race condition (#84)
We've observed that there have been a number of occasions where a newly imported work item could be imported multiple times. It appeared that while the first triggering action was running, a person was continuing to edit the issue. That caused the action to be launched again. Because the first run hadn't finished, multiple new Azure DevOps items were created.

In this PR, I refactored the workflow to avoid that.

Now, the first thing that happens when the request label is added removes that label. That last mutation adds the linked label. That would trigger a second run, but the second run will now always perform an update.