Fix an error with snippets 5000 in CI builds
The CI builds in the API docs project produce PRs that may contain thousands of files.

When the PR contains that many files, the query to retrieve the paths of all the files in the PR throws an invalid operation exception.

Ignore it and return an empty sequence of files.