Correctly update Assignee and current iteration
There's one bug fix and one feature in this PR.

First, the Patch document isn't formatted correctly when un-assigning a Quest issue. The patch document needs to use the "Remove" operation.

Second, update the iteration on the following events:
1. Adding a new Quest item. If it is assigned, put it in the current iteration.
2. When a linked issue is opened / closed, put the associated quest item in the current iteration.

That code uses new query that retrieves all the current iterations when the app starts.