Update story points and iteration in Quest (#188)
* Updates to reading the story points

The docs-maui repo uses one sprint project for the year, and a field for the month. That means it must be read in a different way.

* Initial testing done.

* Add Closing PR in a second request

If the GitHub repo isn't configured as a linked repository for Quest, the closing PR can't be added.

Therefore, update the closing PR in a second REST request, so that all other updates happen correctly when this fails.

* For bulk runs, cache linked status

If the closing PR REST call succeeds, continue to try. If it fails, don't keep sending invalid requests

* add explanatory comment

* Update actions/sequester/Quest2GitHub/AzDoClientServices/JsonPatchDocument.cs

Co-authored-by: David Pine <david.pine@microsoft.com>

---------

Co-authored-by: David Pine <david.pine@microsoft.com>