Clarify the use of `navigationOptions`
The navigation options node in the config is used only when the `--savefile` command line argument isn't specified.

Repos that use a single what's new file, where the newest section is added and the oldest section is removed, don't use the navigation options, but specify the `--savefile` on the command line.