Handle missing child arrays (#279)
Bad news: My last PR broke it.

Good news: Adding tests is paying off.

In all my interactive tests, the `timelineItems` node existed, even if it was empty. Oddly enough, when the SeQuester app runs, some of the issues return with that node missing. I'm still not sure why. The code adds a check that any child array exists and is an array. If not, an empty array is returned.

The added unit test verifies it.