Move slipped items to the backlog iteration
Contributes to #368

When an issue / work item pair is only assigned to iterations in the past, assume it has been slipped. Move that item to the `Future` iteration.

That requires using a new API to retrieve iterations. Instead of retrieve active iterations, retrieve all classification nodes, and parse that packet to build all iterations.  That API includes "parent" or "container" iteration, so filter those out.

The classification nodes API has a different path structure for the iteration nodes, including the "Iteration" classifier. Therefore, remove that element from the path for use assigning the work item to the iteration.