Remove check for `hasChildren`
The current iteration has the flag for "hasChildren" set to true, even though it has no children.

But, because that flag was set, it was filtered out of the possible values for an iteration.

This fix does mean some very old values are included in the list of "all iterations". Those don't match any of the current patterns, so I chose the smallest code change possible. This logic could be modified to remove iterations from old patterns, but that seems unnecessary at this time.