Use System.CommandLine instead of DragonFruit (#609)
* Modernize IssueCloser

IssueCloser uses System.CommandLine, and uses top-level statements.

* factor out command line.

* Convert Snippets5000

Replace DragonFruit with System.CommandLine.

I chose not to convert the Program class to top-level statements because of the number of private static members in this class.

* Modernize Sequester

Replace DragonFruit with System.CommandLine.

While here, convert the main program to top level statements.

* Testing

* Test sample command line parsing